Camino de Vendedor

+++ dominio/Usuario.cs
@@ -13,7 +13,7 @@ namespace dominio
         public string password { get; set; }
         public string nombre { get; set; }
         public string apellido { get; set; }
-        public int dni { get; set; }
+        public long dni { get; set; }
         public string telefono { get; set; }
         public string mail { get; set; }
     }
